Model refactor
Properties of class correct to best understanding and comments added

+++ Deepgram/Records/PreRecorded/Hit.cs
@@ -2,15 +2,27 @@
 
 public record Hit
 {
+    /// <summary>
+    /// Value between 0 and 1 that indicates the model's relative confidence in this hit.
+    /// </summary>
     [JsonPropertyName("confidence")]
-    public double Confidence { get; set; }
+    public decimal Confidence { get; set; }
 
+    /// <summary>
+    /// Offset in seconds from the start of the audio to where the hit occurs.
+    /// </summary>
     [JsonPropertyName("start")]
-    public double Start { get; set; }
+    public decimal Start { get; set; }
 
+    /// <summary>
+    /// Offset in seconds from the start of the audio to where the hit ends.
+    /// </summary>
     [JsonPropertyName("end")]
-    public double End { get; set; }
+    public decimal End { get; set; }
 
+    /// <summary>
+    /// Transcript that corresponds to the time between start and end.
+    /// </summary>
     [JsonPropertyName("snippet")]
     public string Snippet { get; set; }
 }
